mobile::config

Module questions

Structs

Question

Struct for a question The answer here can represent a pure answer or a return for an undone bet

QuestionAndCategory

Struct for a question The answer here can represent a pure answer or a return for an undone bet Added a category for this one for futur usage

QuestionDeck

A deck of question used to get a new question And ensure we don’t repeat already answered one before everything is used.

QuestionDeckIterator

Iterator for a QuestionDeck